hard/requirements.txt
requests
pytelegrambotapi
python-decouple
beautifulsoup4
datetime